+++ betarun/C/sparcdep.h
@@ -10,8 +10,8 @@
 #define _SPARC_H_ 1
 
 struct RegWin {
-    int l0, l1, l2, l3, l4, l5, l6, l7;
-    int i0, i1, i2, i3, i4, i5, fp, i7;
+    long l0, l1, l2, l3, l4, l5, l6, l7;
+    long i0, i1, i2, i3, i4, i5, fp, i7;
 };
 
 /* Heap:
